[Composer 2] Fold processDemoNode into processDocNode as extractDemoSources
